added task and event popup and fixed dropdown box
