+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@react-native-community/slider": "^4.2.4",
+    "@react-native-picker/picker": "2.4.0",
     "expo": "^45.0.6",
     "expo-app-loading": "~2.0.0",
     "expo-font": "~10.1.0",
