+++ solala-app/components/TaskPopup.js
@@ -1,0 +1,18 @@
+import React from "react";
+import AddCalendarPopup from "./CalendarPopup";
+
+const TaskPopup = (props) => {
+    return (
+        <AddCalendarPopup
+            isModalVisible={props.isModalVisible}
+            displayRepeat={true}
+            displayPriority={true}
+            displayCategory={true}
+            displayComplexity={true}
+            displayNotes={true}
+        />
+    );
+
+}
+
+export default TaskPopup;
